refactor(scripts): extract generic hcl and workflow-yaml cores into scripts/lib/ (P7)
Move the HCL tokenizer to scripts/lib/hcl.mjs unchanged and extract the six
generic Actions-workflow helpers into scripts/lib/workflow-yaml.mjs. Five
files outside production-infra-identity-contract/ already read the parser,
including the two ADR 0053 deploy-staging contract files.

Split peg-policy.mjs (1,348 lines) into constants, bucket, publication, and
runtime modules. index.mjs stays the single facade. Every moved line is
byte-identical to its source.

Add scripts/lib/hcl.mjs and scripts/lib/workflow-yaml.mjs to the three
terraform paths-filters and scripts/lib/*.mjs to the quality gate's
contract-surface arm; routing.test.mjs now asserts all four.
