+++ .github/workflows/infra.yml
@@ -23,6 +23,9 @@ on:
       - scripts/check-peg-policy-publication.mjs
       - scripts/check-peg-policy-publication.test.mjs
       - scripts/production-infra-identity-contract/**
+      # Shared parsing cores (ADR 0064) that sit outside the prefix above.
+      - scripts/lib/hcl.mjs
+      - scripts/lib/workflow-yaml.mjs
       - scripts/terraform-fmt-check.mjs
       - scripts/tf-platform-plan-guard.mjs
       - scripts/tf-stacks.mjs
@@ -52,6 +55,9 @@ on:
       - scripts/check-peg-policy-publication.mjs
       - scripts/check-peg-policy-publication.test.mjs
       - scripts/production-infra-identity-contract/**
+      # Shared parsing cores (ADR 0064) that sit outside the prefix above.
+      - scripts/lib/hcl.mjs
+      - scripts/lib/workflow-yaml.mjs
       - scripts/terraform-fmt-check.mjs
       - scripts/tf-platform-plan-guard.mjs
       - scripts/tf-stacks.mjs
