+++ docs/adr/0061-exact-plan-guard-for-manual-platform-applies.md
@@ -120,7 +120,11 @@ retained protected-stack mutation.
 ## Evidence
 
 - Plan policy:
-  [`scripts/check-metrics-bridge-template-plan.mjs`](../../scripts/check-metrics-bridge-template-plan.mjs)
+  [`scripts/check-metrics-bridge-template-plan.mjs`](../../scripts/check-metrics-bridge-template-plan.mjs),
+  which reads HCL through the shared
+  [`scripts/lib/hcl.mjs`](../../scripts/lib/hcl.mjs) core
+  ([ADR 0064](0064-scripts-module-directories.md)). The core carries no policy;
+  the target and refresh rules live here.
 - Wrapper and private-plan lifecycle:
   [`scripts/tf-platform-plan-guard.mjs`](../../scripts/tf-platform-plan-guard.mjs)
   and [`scripts/tf-stacks.mjs`](../../scripts/tf-stacks.mjs)
