+++ scripts/lib/hcl.mjs
